requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
websockets==12.0
aioconsole==0.6.2
prompt-toolkit==3.0.43
